更改通讯协议为 2字节的SID+2字节自身节点的TimeOut+数据。

+++ TestEmbedXrpcForQt/Server/StudentService.Server.h
@@ -11,13 +11,13 @@ BroadcastDataTimeDelegate(EmbedXrpcServerObject *rpcobj)
 {
 this->RpcServerObject=rpcobj;
 }
-uint32_t GetSid(){return BroadcastDataTime_ServiceId;}
+uint16_t GetSid(){return BroadcastDataTime_ServiceId;}
 void  Invoke(DateTime_t t);
 };
 class GetStudentInfoFormStudentIdService:public IService
 {
 public:
-uint32_t GetSid(){return GetStudentInfoFormStudentId_ServiceId;}
+uint16_t GetSid(){return GetStudentInfoFormStudentId_ServiceId;}
 GetStudentInfoFormStudentId_Response Response;
 void GetStudentInfoFormStudentId(Byte StudentIdLen,Byte StudentId[100],Int32 arg2,Int32 arg3);
 void Invoke(SerializationManager &recManager, SerializationManager& sendManager);
@@ -25,7 +25,7 @@ void Invoke(SerializationManager &recManager, SerializationManager& sendManager)
 class GetStudentsInfoFormAgeService:public IService
 {
 public:
-uint32_t GetSid(){return GetStudentsInfoFormAge_ServiceId;}
+uint16_t GetSid(){return GetStudentsInfoFormAge_ServiceId;}
 GetStudentsInfoFormAge_Response Response;
 void GetStudentsInfoFormAge();
 void Invoke(SerializationManager &recManager, SerializationManager& sendManager);
@@ -33,7 +33,7 @@ void Invoke(SerializationManager &recManager, SerializationManager& sendManager)
 class TestService:public IService
 {
 public:
-uint32_t GetSid(){return Test_ServiceId;}
+uint16_t GetSid(){return Test_ServiceId;}
 void Test();
 void Invoke(SerializationManager &recManager, SerializationManager& sendManager);
 };
